Added the first tool (written in Microsoft[R] PowerShell)
This commit defines a tool which lists all checked out files in all
folders accessible by a particular user. For this script, it is required
to have access to the administrative site as well. It can however quite
easily be refurbished to use only a particular site (for example in a
project environment)